requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113 
torch
torchvision
git+https://github.com/huggingface/diffusers.git
transformers
ftfy
numpy
matplotlib
uuid
opencv-python
git+https://github.com/openai/CLIP.git
python-dotenv
gradio